requirements.txt
streamlit>=1.22.0
pandas>=2.0
altair>=5.0
openpyxl>=3.1
python-docx>=0.8.12
pdfplumber>=0.7
